edit cart

+++ internal/models/category.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Category struct {
 	gorm.Model
-	ID          int64  `gorm:"primary_key"`
+	ID          int    `gorm:"primary_key"`
 	Name        string `gorm:"unique"`
 	Description string
 	Products    []Product
