+++ internal/models/contact.go
@@ -6,7 +6,7 @@ import (
 
 type Contact struct {
 	gorm.Model
-	ID      int64  `gorm:"primary_key"`
+	ID      int    `gorm:"primary_key"`
 	Name    string `json:"name"`
 	Email   string `json:"email"`
 	Message string `json:"message"`
