+++ internal/models/product.go
@@ -4,12 +4,12 @@ import "gorm.io/gorm"
 
 type Product struct {
 	gorm.Model
-	ID          int64  `gorm:"primary_key"`
+	ID          int    `gorm:"primary_key"`
 	Name        string `gorm:"unique"`
 	Description string
 	Price       float64
-	Stock       int64
-	CategoryID  int64
+	Stock       int
+	CategoryID  int
 	Category    Category
 	ImageURL    string
 }
